Fix and simplify the stopping of WorkerThread instances.
WorkerThread.stopRunning was guaranteed to throw an exception if the thread instance had not already stopped running. Worse, the exception would occur before setting the crucial "running" flag to false. The exception was generated by calling this.wait() without owning a lock on this.

Minimally, the call to this.wait() needed to be removed and the setting of the "running" flag to false needed to be moved to the top of the method. However, I could find no purpose for the method to do anything but set the "running" flag. The wait and interrupt calls seem to be meaningless detours. Therefore, I trimmed the method to simply set the flag and return, which is working well in practice.