Support for new "Templates" projects
This commit adds support for the split “Templates” projects
(SDL/SDLECOM) by introducing a new build property which tells the Ant
built which template project to pull from.

The Travis build is updated by this commit to handle the new property,
the Jenkins builds need to be updated now too, which I’ll do next.